Updated client.js and server.js

+++ client.js
@@ -2,6 +2,9 @@
 const net = require('net');
 const fs = require('fs');
 const port = 8124;
+const firstRequestStr = 'FILES';
+const successReq = 'ASC';
+const failedReq = 'DEC';
 
 const client = new net.Socket();
 
@@ -9,12 +12,16 @@ client.setEncoding('utf8');
 
 client.connect(port, function() {
   console.log('Connected');
-  client.write('\r\nHello, Server!\r\nLove,\r\nClient.\r\n');
+  client.write(firstRequestStr);
 });
 
 client.on('data', function(data) {
   console.log(data);
-  client.destroy();
+
+  if (data == firstRequestStr){
+  }else{
+    client.destroy();
+  }
 });
 
 client.on('close', function() {
